# [IMP] core: allow replacing COALESCE with UNION in sql quieries