UNION ALL is usually better than single SELECT with COALESCE, because it allows
postgres to make a better query plan.

New feature is applied for translation field, but could be used in other cases
too.

---

task-2081307

Description of the issue/feature this PR addresses:

Current behavior before PR:

Desired behavior after PR is merged:




--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr