@Yenthe666 2021-02-15
Hmm interesting. Any benchmarks/perf tests?