@yelizariev 2021-02-17
> Hmm interesting. Any benchmarks/perf tests?

I tested with search product.template by name:

For small databases it doesn't make it worse.:

before: 3,5 ms
after: 1,9 ms

For a database populated with dummy values (700 K products, 1.6 M products) and running on postgres v12:

before: 700 ms
after: 70 ms

With real usage the improvement should give greater speed up, because of bigger entropy, using postgres v10 (like in odoo.sh :stuck_out_tongue_winking_eye: ) and the fact that name_search for products makes several requests on translated fields