@yelizariev 2021-03-22
It kind of works, but does't give a speed up, but make it worse in /shop :cry: 

I'll try later with bigger database and after /shop improvements in https://github.com/odoo/odoo/pull/68113

# Performance test

* Odoo 15 (master)
* Postgres 12.5
* Demo database with website_sale installed
* 87K product.templates with 4 variants in each
* English and French language active
* Egnlish is current language
* Numbers in the table are
 * Number of queries
 * Query time, sec
 * Remaining time, sec

```
|                       | product.product search by name (few results) | product.template search by name (few results) | /shop without search | /shop with search (few results) | /shop with search (99% in results) |   |   |
|-----------------------+----------------------------------------------+-----------------------------------------------+----------------------+---------------------------------+------------------------------------+---+---|
| Before                | 12 0.340 0.040                               | 187 0.180 0.097                               | 245 0.341 0.179      | 259 0.420 0.179                 | 259 0.738 0.201                    |   |   |
| After (MAX_UNIONS=16) | 12 3.540 0.038                               | 187 0.211 0.106                               | 243 0.354 0.181      | 259 2.411 0.163                 | 259 2.810 0.227                    |   |   |
| After (MAX_UNIONS=2)  | 12 0.235 0.044                               | 187 0.205 0.109                               | 245 0.339 0.193      | 259 0.621 0.179                 | 259 1.152 0.245                    |   |   |
```