@yelizariev 2023-02-13
Too complicated, quite obsolete (v16+ doesn't have ir_translation table)